+++ circle.h
@@ -19,7 +19,7 @@ class Circle : public Shape
 
 public:
 	Circle();
-	Circle(const Point&, int);
+	Circle(const Point&, double);
 	Circle(const Circle&);			// 깊은 복사를 할 필요가 없다면 프로그래머가 복사생성자를
 									// 프로그램할 필요는 없다
 									// 만약 어떤 이유로 프로그램해야 한다면 메모리를 이해하고
